refactor: unify expression directory structure
- Remove expressions.py, use expression/ directory instead
- Add expression/__init__.py as entry point
- Split definitions into expression/load_data.py, json_table.py
- Update test file import paths
