+++ src/rhosocial/activerecord/backend/impl/mysql/expression/__init__.py
@@ -1,0 +1,22 @@
+# src/rhosocial/activerecord/backend/impl/mysql/expression/__init__.py
+"""
+MySQL-specific expression classes.
+
+This module provides expression classes that are specific to MySQL,
+such as LOAD DATA INFILE and JSON_TABLE expressions.
+
+Directory structure:
+- load_data.py    - LOAD DATA INFILE expression
+- json_table.py   - JSON_TABLE expression
+"""
+
+from .load_data import LoadDataExpression, LoadDataOptions
+from .json_table import JSONTableExpression, JSONTableColumn, NestedPath
+
+__all__ = [
+    "LoadDataExpression",
+    "LoadDataOptions",
+    "JSONTableExpression",
+    "JSONTableColumn",
+    "NestedPath",
+]
