+++ tests/rhosocial/activerecord_mysql_test/feature/query/test_json_table.py
@@ -11,7 +11,7 @@ Official Documentation:
 import pytest
 import pytest_asyncio
 
-from rhosocial.activerecord.backend.impl.mysql.expressions import (
+from rhosocial.activerecord.backend.impl.mysql.expression import (
     JSONTableExpression, JSONTableColumn, NestedPath
 )
